Christmas Greeting Animation Addon for Blender
This add-on is for creating greeting cards for Christmas in Blender. You can read the names from a CSV file and select the elements and numbers that you want to add to the animation.

Documentation:

Delete all objects in our first scene with our panel.
Click the button to generate snowflakes if we want to use them in the animation in our second scene.
If there is a CSV file that we want to read, we write its path and click the Read button. The names will appear in the terminal window.
We write our greeting or message in the Message section and click the Add Text button.
In the "To Who" section, we write the name of the person we want to send the message to and click the Add Text button.
We enter the number of spheres we want to use as fireworks in the scene in the "Number of Spheres" section. After the sphere count reaches 45, it becomes more visually appealing. After entering the number, we click the Create Animated Spheres button.
When we click the Generate Scene button, we add the light, background, camera and lens settings, frame intervals, render shading type, and other settings we use in the scene.
We enter the number of trees we want to add to the scene in the panel where we specify the number of trees. To create a visually pleasing scene, it is better to create more than 40 trees. After entering the number, we click the Add Trees button.
We can see how our scene will look when we toggle the camera view or press the Numpad 0 key, and we can watch our animation by clicking the play button.
We can render the animation with the base render settings in our scene by pressing Command + F12 keys.
Translation: I have created a Blender Python add-on code in a Github repository that enables users to create Christmas greeting cards in Blender by reading names from a CSV file and selecting elements and numbers for their animation. To install the add-on in Blender, follow these steps:

In Blender, go to Edit > Preferences > Add-ons. Click on the Install button at the top right of the Preferences window. Navigate to the location of the downloaded add-on and select it. Click the Install Add-on button to install it.

Once installed, you can use the add-on with the following steps:

Delete all objects in the first scene using the panel. Generate snowflakes by clicking the button in the second scene, if desired. If you have a CSV file to read, enter its path and click the Read button. Names will appear in the terminal window. Write your greeting or message in the Message section and click the Add Text button. Enter the name of the person you want to send the message to in the "To Who" section and click the Add Text button. Enter the number of spheres you want to use as fireworks in the "Number of Spheres" section. After 45 spheres, the visual appeal increases. Click the Create Animated Spheres button. Click the Generate Scene button to add light, background, camera and lens settings, frame intervals, render shading type, and other settings to your scene. Enter the number of trees you want to add to the scene in the panel. Creating more than 40 trees

This addon had tried in Blender 3.3.1 version.